+++ dflib/src/test/java/org/dflib/print/TabularPrinter_DataFrameTest.java
@@ -2,34 +2,31 @@ package org.dflib.print;
 
 import org.dflib.DataFrame;
 import org.dflib.Series;
-import org.dflib.print.TabularPrinter;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TabularPrinter_DataFrameTest {
 
-    private DataFrame df;
+    private static final String LS = System.lineSeparator();
 
-    @BeforeEach
-    public void initDataFrame() {
-        this.df = DataFrame.byColumn("col1", "column2").of(
-                Series.of("one", "two", "three", "four"),
-                Series.ofInt(1, 2, 3, 44));
-    }
+    private static final DataFrame df = DataFrame
+            .byColumn("col1", "column2")
+            .of(
+                    Series.of("one", "two", "three", "four"),
+                    Series.ofInt(1, 2, 3, 44));
 
     @Test
     public void toString_Normal() {
         TabularPrinter p = new TabularPrinter(5, 10);
 
-        assertEquals(System.lineSeparator() +
-                "col1  column2" + System.lineSeparator() +
-                "----- -------" + System.lineSeparator() +
-                "one         1" + System.lineSeparator() +
-                "two         2" + System.lineSeparator() +
-                "three       3" + System.lineSeparator() +
-                "four       44" + System.lineSeparator() +
+        assertEquals(LS +
+                "col1  column2" + LS +
+                "----- -------" + LS +
+                "one         1" + LS +
+                "two         2" + LS +
+                "three       3" + LS +
+                "four       44" + LS +
                 "4 rows x 2 columns", p.toString(df));
     }
 
@@ -37,13 +34,13 @@ public class TabularPrinter_DataFrameTest {
     public void toString_Name() {
         TabularPrinter p = new TabularPrinter(5, 10);
 
-        assertEquals(System.lineSeparator() +
-                "col1  column2" + System.lineSeparator() +
-                "----- -------" + System.lineSeparator() +
-                "one         1" + System.lineSeparator() +
-                "two         2" + System.lineSeparator() +
-                "three       3" + System.lineSeparator() +
-                "four       44" + System.lineSeparator() +
+        assertEquals(LS +
+                "col1  column2" + LS +
+                "----- -------" + LS +
+                "one         1" + LS +
+                "two         2" + LS +
+                "three       3" + LS +
+                "four       44" + LS +
                 "[df0] 4 rows x 2 columns", p.toString(df.as("df0")));
     }
 
@@ -51,12 +48,12 @@ public class TabularPrinter_DataFrameTest {
     public void toString_TruncateRows() {
         TabularPrinter p = new TabularPrinter(2, 10);
 
-        assertEquals(System.lineSeparator() +
-                "col1 column2" + System.lineSeparator() +
-                "---- -------" + System.lineSeparator() +
-                "one        1" + System.lineSeparator() +
-                "..." + System.lineSeparator() +
-                "four      44" + System.lineSeparator() +
+        assertEquals(LS +
+                "col1 column2" + LS +
+                "---- -------" + LS +
+                "one        1" + LS +
+                "..." + LS +
+                "four      44" + LS +
                 "4 rows x 2 columns", p.toString(df));
     }
 
@@ -64,13 +61,32 @@ public class TabularPrinter_DataFrameTest {
     public void toString_TruncateColumns() {
         TabularPrinter p = new TabularPrinter(5, 4);
 
-        assertEquals(System.lineSeparator() +
-                "col1 c..2" + System.lineSeparator() +
-                "---- ----" + System.lineSeparator() +
-                "one     1" + System.lineSeparator() +
-                "two     2" + System.lineSeparator() +
-                "t..e    3" + System.lineSeparator() +
-                "four   44" + System.lineSeparator() +
+        assertEquals(LS +
+                "col1 c..2" + LS +
+                "---- ----" + LS +
+                "one     1" + LS +
+                "two     2" + LS +
+                "t..e    3" + LS +
+                "four   44" + LS +
                 "4 rows x 2 columns", p.toString(df));
+    }
+
+    @Test
+    public void emptyColumn() {
+
+        DataFrame df = DataFrame
+                .byColumn("col1", "")
+                .of(
+                        Series.of("one", "two"),
+                        Series.of("", ""));
+
+        TabularPrinter p = new TabularPrinter(5, 10);
+
+        assertEquals(LS +
+                "col1  " + LS +
+                "---- -" + LS +
+                "one   " + LS +
+                "two   " + LS +
+                "[df0] 2 rows x 2 columns", p.toString(df.as("df0")));
     }
 }