+++ RELEASE-NOTES.md
@@ -15,6 +15,7 @@
 * #360 RowColumnSet: rename "map" to "merge"
 * #364 Empty series isTrue(..) should return "false"
 * #365 BooleanSeries.eq(..) : inefficiency and potential ClassCastException
+* #375 Exception printing empty columns
 
 ## Release 1.0.0-RC1